Add files via upload
This repository contains the UI/UX design work for a Food Delivery App, developed as part of my internship. The project aims to create an intuitive and seamless user experience for ordering food online.
The design boasts a user-friendly interface with a clean and simple design that enhances usability. It features smooth navigation with easy-to-navigate menus and categories for a seamless browsing experience. The design includes interactive elements that engage users and improve interaction. It is also optimized for responsiveness, ensuring a consistent experience on both mobile and tablet devices.