Add files via upload
This repository contains the design and implementation of sign-up and log-in forms as part of my internship project in the domain of UI/UX. The forms are designed to provide a seamless and user-friendly experience for users accessing a website.
The forms boast a responsive design, ensuring accessibility and a great appearance across all devices, including desktops, tablets, and mobile phones. They feature a user-friendly interface with a clear and intuitive layout, utilizing appropriate form elements to enhance user experience. Client-side validation guides users in providing the correct information, while the modern and clean design focuses on attention to detail, consistent styling, and cohesive color schemes.
You can use these forms as a starting point for your own projects or integrate them into an existing website. Feel free to customize the design and functionality to fit your specific needs.
Contributions are welcome! If you have suggestions for improvements or have found any bugs, please open an issue or submit a pull request.