Add files via upload
This repository contains the UI/UX design for an e-commerce website, created as part of my internship. The design focuses on enhancing user experience and usability while providing a visually appealing interface.
The design includes a welcoming and informative Home Page featuring products, promotions, and navigation. The Product Listing page is organized into categories with filters and sorting options for easy browsing. Detailed product information, images, reviews, and related products are provided on the Product Detail page. The Cart and Checkout process is user-friendly and seamless, enhancing the overall shopping experience. The User Account section allows for profile management, and wishlist functionalities. Additionally, the design is fully responsive, ensuring optimal performance on both desktop and mobile devices.
Guidelines for contributing to the project, if you are open to collaboration.